+++ vcd/datasource_vcd_catalog_item.go
@@ -1,0 +1,33 @@
+package vcd
+
+import "github.com/hashicorp/terraform/helper/schema"
+
+func datasourceVcdCatalogItem() *schema.Resource {
+	return &schema.Resource{
+		Read: resourceVcdCatalogItemRead,
+		Schema: map[string]*schema.Schema{
+			"org": {
+				Type:     schema.TypeString,
+				Optional: true,
+			},
+			"catalog": &schema.Schema{
+				Type:        schema.TypeString,
+				Required:    true,
+				Description: "catalog name where upload the OVA file",
+			},
+			"name": &schema.Schema{
+				Type:     schema.TypeString,
+				Required: true,
+			},
+			"description": &schema.Schema{
+				Type:     schema.TypeString,
+				Computed: true,
+			},
+			"metadata": {
+				Type:        schema.TypeMap,
+				Computed:    true,
+				Description: "Key and value pairs for catalog item metadata",
+			},
+		},
+	}
+}