+++ CHANGELOG.md
@@ -3,10 +3,14 @@
 FEATURES:
 
 * **New Data Source:** `vcd_org` Organization - ([#218](https://github.com/terraform-providers/terraform-provider-vcd/issues/218))
+* **New Data Source:** `vcd_catalog` Catalog - ([#218](https://github.com/terraform-providers/terraform-provider-vcd/issues/218))
+* **New Data Source:** `vcd_catalog_item` CatalogItem - ([#218](https://github.com/terraform-providers/terraform-provider-vcd/issues/218))
 
 IMPROVEMENTS:
 
-* `vcd_org` Add import capability
+* `resource/vcd_org` Add import capability
+* `resource/vcd_catalog` Add import capability
+* `resource/vcd_catalog_item` Add import capability
 * `resource/catalog_item` added catalog item metadata support [#28*] 
 * `resource/catalog_media` added catalog media item metadata support [#28*]